Fleshing out more of the bundler chapter.
It would appear that I'm writing two things: instructions and
commentary. My goal is for the instructions to be referenceable and
with minimal explanation be regarded as important.

The commentary, in the footnotes, is something helps strengthen or
explain why I chose to include a particular command.

By separating instruction and commentary, I'm hoping it is easier to
ultimatley construct the final form of the book. While I may read
footnotes, I don't know how customary that is. But I don't want to
wax philosophical in the instructions.